refactor(parts): port remainder of json to yaml
