+++ packages/parts/src/data/segment-lookup.json
