+++ packages/parts/src/info.ts
@@ -1,4 +1,4 @@
-import SEGMENT_LOOKUP_SOURCE from './segment-lookup.json' with { type: 'json' }
+import SEGMENT_LOOKUP_SOURCE from './data/segment-lookup.json' with { type: 'json' }
 import SPRITE_DEFS from './data/sprite_defs.json' with { type: 'json' }
 import {
   // getSegmentLookup,
